Refactor chunks into separate package
