+++ src/main/java/com/gorman/voxel_engine/world/terrain/FlatTerrain.java
@@ -1,5 +1,6 @@
 package com.gorman.voxel_engine.world.terrain;
 
+import com.gorman.voxel_engine.world.chunks.Chunk;
 import com.gorman.voxel_engine.world.primitives.Vector;
 import com.gorman.voxel_engine.world.voxels.Grass;
 import com.gorman.voxel_engine.world.voxels.Stone;
