+++ src/main/java/com/gorman/voxel_engine/world/terrain/Terrain.java
@@ -1,5 +1,6 @@
 package com.gorman.voxel_engine.world.terrain;
 
+import com.gorman.voxel_engine.world.chunks.Chunk;
 import com.gorman.voxel_engine.world.primitives.Vector;
 
 public abstract class Terrain {
