+++ src/test/java/com/gorman/voxel_engine/world/chunks/TestChunk.java
@@ -1,4 +1,4 @@
-package com.gorman.voxel_engine.world.terrain;
+package com.gorman.voxel_engine.world.chunks;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
